backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.41
asyncpg==0.31.0
alembic==1.16.1
pydantic==2.12.2
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
httpx==0.28.1
redis[hiredis]==5.3.0
cryptography==45.0.3
python-multipart==0.0.20
stripe==12.2.0
